requirements.txt
absl-py==0.7.1
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
graphviz==0.11
grpcio==1.21.1
h5py==2.9.0
idna==2.8
igraph==0.1.11
imageio==2.5.0
importlib-metadata==0.18
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.0
jgraph==0.2.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.29.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numba==0.44.1
numpy==1.16.4
packaging==19.0
pandapower==2.0.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pdf2image==1.9.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
py==1.8.0
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.3
pytest==4.6.3
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.0.2
qtconsole==4.5.2
requests==2.23.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
simbench==1.0.0
six==1.12.0
sklearn==0.0
sxl==0.0.1a3
sympy==1.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
zipp==0.5.1
